+++ src/Pim/Bundle/EnrichBundle/MassEditAction/Operation/EditCommonAttributes.php
@@ -59,6 +59,15 @@ class EditCommonAttributes extends ProductMassEditOperation
     /** @var ProductBuilder */
     protected $productBuilder;
 
+    /** @var string */
+    protected $productPriceClass;
+
+    /** @var string */
+    protected $productMediaClass;
+
+    /** @var string */
+    protected $metricClass;
+
     /**
      * Constructor
      *
@@ -68,6 +77,7 @@ class EditCommonAttributes extends ProductMassEditOperation
      * @param CatalogContext           $catalogContext
      * @param ProductBuilder           $productBuilder
      * @param ProductMassActionManager $massActionManager
+     * @param array                    $classes
      */
     public function __construct(
         ProductManager $productManager,
@@ -75,7 +85,8 @@ class EditCommonAttributes extends ProductMassEditOperation
         CurrencyManager $currencyManager,
         CatalogContext $catalogContext,
         ProductBuilder $productBuilder,
-        ProductMassActionManager $massActionManager
+        ProductMassActionManager $massActionManager,
+        array $classes
     ) {
         $this->productManager = $productManager;
         $this->userContext = $userContext;
@@ -85,6 +96,9 @@ class EditCommonAttributes extends ProductMassEditOperation
         $this->massActionManager = $massActionManager;
         $this->displayedAttributes = new ArrayCollection();
         $this->values = new ArrayCollection();
+        $this->productPriceClass = $classes['product_price'];
+        $this->productMediaClass = $classes['product_media'];
+        $this->metricClass = $classes['metric'];
     }
 
     /**
@@ -406,7 +420,7 @@ class EditCommonAttributes extends ProductMassEditOperation
      */
     protected function createProductPrice($currency)
     {
-        return new ProductPrice(null, $currency);
+        return new $this->productPriceClass(null, $currency);
     }
 
     /**
@@ -453,7 +467,7 @@ class EditCommonAttributes extends ProductMassEditOperation
     protected function setProductFile(ProductValueInterface $productValue, ProductValueInterface $value)
     {
         if (null === $media = $productValue->getMedia()) {
-            $media = new ProductMedia();
+            $media = new $this->productMediaClass();
             $productValue->setMedia($media);
         }
         $file = $value->getMedia()->getFile();
@@ -471,7 +485,7 @@ class EditCommonAttributes extends ProductMassEditOperation
     protected function setProductMetric(ProductValueInterface $productValue, ProductValueInterface $value)
     {
         if (null === $metric = $productValue->getMetric()) {
-            $metric = new Metric();
+            $metric = new $this->metricClass();
             $metric->setFamily($value->getAttribute()->getMetricFamily());
             $productValue->setMetric($metric);
         }